+++ hutool-core/src/main/java/cn/hutool/core/date/DateUtil.java
@@ -659,7 +659,7 @@ public class DateUtil {
 	 */
 	public static DateTime parseTimeToday(String timeString) {
 		timeString = StrUtil.format("{} {}", today(), timeString);
-		return parse(timeString, DatePattern.NORM_TIME_FORMAT);
+		return parse(timeString, DatePattern.NORM_DATETIME_PATTERN);
 	}
 
 	/**