+++ hutool-http/src/main/java/cn/hutool/http/HttpRequest.java
@@ -101,7 +101,12 @@ public class HttpRequest extends HttpBase<HttpRequest> {
 	 * @return HttpRequest
 	 */
 	public HttpRequest method(Method method) {
-		this.method = method;
+		if(Method.PATCH == method) {
+			this.method = Method.POST;
+			this.header("X-HTTP-Method-Override", "PATCH");
+		}else {
+			this.method = method;
+		}
 		return this;
 	}